Moving implementation of sorted set in new package set. Removing ObjectNotFound & DuplicateElement exceptions. Adding implementation of hashCode().
