--- MySortedSet/homework4/exceptions/DuplicateElementException.java
@@ -1,7 +1,0 @@
-package homework4.exceptions;
-
-public class DuplicateElementException extends RuntimeException {
-    public DuplicateElementException(String message) {
-        super(message);
-    }
-}
