--- MySortedSet/homework4/exceptions/ObjectNotFoundException.java
@@ -1,7 +1,0 @@
-package homework4.exceptions;
-
-public class ObjectNotFoundException extends RuntimeException {
-    public ObjectNotFoundException(String message) {
-        super(message);
-    }
-}
