+++ MySortedSet/homework4/set/MyBinaryTree.java
@@ -1,4 +1,4 @@
-package homework4.model;
+package homework4.set;
 
 /**
  * Custom implementation of a Binary Search Tree.
