+++ MySortedSet/homework4/test/GlassesSetTest.java
@@ -1,7 +1,7 @@
 package homework4.test;
 
 import homework4.model.Glasses;
-import homework4.model.MySortedSet;
+import homework4.set.MySortedSet;
 
 public class GlassesSetTest {
     public static void main(String[] args) {
